refactor: extract use-contact-tag-action.tsx - remove empty files - extract vCard export hook - extract use-apply-tag-contacts.tsx - extract use-contact-show-tag-action.tsx
