+++ src/legacy/ui-actions/tag-actions.tsx
@@ -11,12 +11,11 @@ import {
 	Padding,
 	Icon,
 	Checkbox,
-	Button,
 	useModal,
 	useSnackbar
 } from '@zextras/carbonio-design-system';
 import { TFunction } from 'i18next';
-import { every, find, includes, map, noop, reduce } from 'lodash';
+import { every, find, includes, map, reduce } from 'lodash';
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
 
@@ -394,82 +393,6 @@ export const applyMultiTag = ({
 	return {
 		id: TagsActionsType.APPLY,
 		items: tagItem,
-		customComponent: (
-			<Row takeAvailableSpace mainAlignment="flex-start">
-				<Padding right="small">
-					<Icon icon="TagsMoreOutline" />
-				</Padding>
-				<Row takeAvailableSpace mainAlignment="space-between">
-					<Padding right="small">
-						<Text>{t('label.tags', 'Tags')}</Text>
-					</Padding>
-				</Row>
-			</Row>
-		)
-	};
-};
-export const applyTag = ({
-	t,
-	contact,
-	tags,
-	createModal,
-	closeModal
-}: {
-	t: TFunction;
-	contact: any;
-	tags: TagsFromStoreType;
-	createModal: ReturnType<typeof useModal>['createModal'];
-	closeModal: ReturnType<typeof useModal>['closeModal'];
-}): {
-	id: string;
-	items: ItemType[];
-	customComponent: ReactElement;
-	label?: string;
-	icon?: string;
-	onClick: () => void;
-} => {
-	const tagItem = reduce(
-		tags,
-		(acc, v) => {
-			const item = {
-				id: v.id,
-				label: v.name,
-				icon: 'TagOutline',
-				keepOpen: true,
-				customComponent: <TagsDropdownItem tag={v} contact={contact} />
-			};
-			// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-			// @ts-ignore
-			acc.push(item);
-			return acc;
-		},
-		[]
-	);
-	const newTag = {
-		id: 'new_tag',
-		keepOpen: true,
-		customComponent: (
-			<Button
-				label={t('label.new_tag', 'New Tag')}
-				type="outlined"
-				width="fill"
-				size="small"
-				onClick={(ev): void => {
-					createAndApplyTag({ t, contact, createModal, closeModal }).onClick?.(ev);
-				}}
-			/>
-		)
-	};
-	// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-	// @ts-ignore
-	tagItem.push(newTag);
-
-	return {
-		id: TagsActionsType.APPLY,
-		items: tagItem,
-		label: t('label.tag', 'Tag'),
-		icon: 'TagsMoreOutline',
-		onClick: noop,
 		customComponent: (
 			<Row takeAvailableSpace mainAlignment="flex-start">
 				<Padding right="small">
